+++ R/fastcca.R
@@ -46,14 +46,17 @@ setMethod("show", "fastcca", function(object) {
 #'
 #' Fast Canonical correlation analysis that is scalable to high dimensional data.  Uses covariance shrinkage and algorithmic speed ups to be linear time in p when p > n.
 #'
-#' @param X first matrix (n x p1)
-#' @param Y first matrix (n x p2)
+#' @param X \code{matrix}, \code{data.frame}, \code{eclairs} decomposition or \code{list} from \code{svd()}
+#' @param Y \code{matrix}, \code{data.frame}, \code{eclairs} decomposition or \code{list} from \code{svd()}
 #' @param k number of canonical components to return
 #' @param k.x number of singular vectors of X to use
 #' @param k.y number of singular vectors of Y to use
 #' @param lambda.x shrinkage parameter for X. If \code{NULL}, estimate from data
 #' @param lambda.y shrinkage parameter for Y. If \code{NULL}, estimate from data
 #' @param svd.method SVD algorithm string "svd", "irlba", or "pcaone" 
+#'
+#' @details
+#' Objects in \code{X} and \code{Y} are converted to \code{eclairs} decomposition from \code{matrix}, \code{data.frame} or \code{list} from \code{svd()}.  The \code{predict()} function maps from one input to another, except when input object is \code{list} from \code{svd()}.
 #'
 #' @examples
 #' pop <- LifeCycleSavings[, 2:3]
@@ -65,10 +68,27 @@ setMethod("show", "fastcca", function(object) {
 #' 
 #' # predict second dataset given first
 #' y.pred <- predict(fit, X = pop)
+#'
+#' # Advanced
+#' # fastcca() also works with eclairs() objects
+#' ecl.x <- eclairs(as.matrix(pop))
+#' ecl.y <- eclairs(as.matrix(oec))
+#' fit2 = fastcca(ecl.x, ecl.y)
+#' 
+#' # fastcca() also works mixed types
+#' fit3 = fastcca(pop, ecl.y)
+#' 
+#' # it also works with results of svd()
 #
 #' @importFrom Rfast cora colsums
 #' @export
-fastcca <- function(X, Y, k = NULL, k.x=min(dim(X)), k.y=min(dim(Y)), lambda.x=NULL, lambda.y=NULL, svd.method = c("svd", "irlba", "pcaone") ){
+fastcca <- function( X, Y, 
+  k = NULL, 
+  k.x = NULL, 
+  k.y = NULL, 
+  lambda.x = NULL, 
+  lambda.y = NULL, 
+  svd.method = c("svd", "irlba", "pcaone") ){
 
   # checks
   stopifnot("k must be positive" = k > 0)
@@ -79,30 +99,116 @@ fastcca <- function(X, Y, k = NULL, k.x=min(dim(X)), k.y=min(dim(Y)), lambda.x=N
   stopifnot("lambda.y must be in (0,1)" = (lambda.y >= 0) & (lambda.y <=1))
   svd.method <- match.arg(svd.method)
 
-  if (!is.matrix(X)) {
-    X <- as.matrix(X)
+  # if (nrow(X) != nrow(Y)) {
+  #   stop("X and Y must have the same number of rows")
+  # }
+
+  # Convert X to eclairs decomposition
+  if( is.matrix(X) | is.data.frame(X) ){
+
+    if( is.null(k.x) ) k.x <- min(dim(X))
+
+    # SVD and shrinkage
+    X_scaled <- .standardise(X)
+    ecl.x <- eclairs(X_scaled, 
+              k = k.x, 
+              lambda = lambda.x, 
+              svd.method = svd.method)
+
+  }else if( is(X, "eclairs")){
+    ecl.x <- X
+  }else if( is(X, "list") ){
+    dcmp <- X;
+    ecl.x <- as.eclairs(dcmp, 
+                k = ncol(dcmp$u), 
+                n = nrow(dcmp$u), 
+                p = ncol(dcmp$v),
+                svd.method = "svd",
+                lambda = lambda.x)
   }
-  if (!is.matrix(Y)) {
-    Y <- as.matrix(Y)
-  }
-  if (nrow(X) != nrow(Y)) {
-    stop("X and Y must have the same number of rows")
+ 
+  # Convert Y to eclairs decomposition
+  if( is.matrix(Y) | is.data.frame(Y) ){
+
+    if( is.null(k.y) ) k.y <- min(dim(Y))
+
+    # SVD and shrinkage
+    Y_scaled <- .standardise(Y)
+    ecl.y <- eclairs(Y_scaled, 
+              k = k.y, 
+              lambda = lambda.y, 
+              svd.method = svd.method)
+
+  }else if( is(Y, "eclairs")){
+    ecl.y <- Y
+  }else if( is(Y, "list") ){
+    dcmp <- Y;
+    ecl.y <- as.eclairs(dcmp, 
+                k = ncol(dcmp$u), 
+                n = nrow(dcmp$u), 
+                p = ncol(dcmp$v),
+                svd.method = "svd",
+                lambda = lambda.y)
   }
 
-  # scale columns
-  X_scaled <- .standardise(X)
-  Y_scaled <- .standardise(Y)
+  # Perform CCA
+  res = .fastcca( ecl.x, ecl.y, k = k, svd.method = svd.method)
 
-  # SVD and shrinkage
-  ecl.x <- eclairs(X_scaled, 
-    k = k.x, 
-    lambda = lambda.x, 
-    svd.method = svd.method)
+  # Assign variable names
+  rownames(res$y.coefs) <- colnames(Y)
 
-  ecl.y <- eclairs(Y_scaled, 
-    k = k.y, 
-    lambda = lambda.y, 
-    svd.method = svd.method)
+  # Compute latent variables
+  if( exists("X_scaled") ){
+    rownames(res$x.coefs) <- colnames(X_scaled)
+    res$x.vars <- X_scaled %*% res$x.coefs
+    res$x.mu = attr(X_scaled, "mu")
+    res$x.sd = attr(X_scaled, "sd")
+  }else{
+    rownames(res$x.coefs) <- ecl.x$colnames
+
+    # X_recon <- with(ecl.x, V %*% diag(sqrt(dSq)) %*% t(U) )
+    # res$x.vars <- X_recon %*% res$x.coef
+    res$x.vars <- dmult(ecl.x$V, sqrt(ecl.x$dSq), "right") %*% 
+      crossprod(ecl.x$U, res$x.coef)
+
+    res$x.mu = ecl.x$mu
+    res$x.sd = ecl.x$sigma
+  }
+  if( exists("Y_scaled") ){
+    rownames(res$y.coefs) <- colnames(Y_scaled)
+    res$y.vars <- Y_scaled %*% res$y.coefs
+    res$y.mu = attr(Y_scaled, "mu")
+    res$y.sd = attr(Y_scaled, "sd")
+  }else{
+    rownames(res$y.coefs) <- ecl.y$colnames
+
+    # Y_recon <- with(ecl.y, V %*% diag(sqrt(dSq)) %*% t(U) )
+    # res$y.vars <- Y_recon %*% res$y.coef
+    res$y.vars <- dmult(ecl.y$V, sqrt(ecl.y$dSq), "right") %*% 
+      crossprod(ecl.y$U, res$y.coef)
+
+    res$y.mu = ecl.y$mu
+    res$y.sd = ecl.y$sigma
+  }
+
+  # Faster way to eval diag of correlation
+  # diag(cor(x.vars, y.vars))[seq(k)]
+  res$cor <- colsums(.standardise(res$x.vars) * .standardise(res$y.vars)) / (nrow(res$y.vars)-1)
+  names(res$cor) <- paste("comp", seq(k), sep = "")
+
+  res
+}
+
+
+
+
+
+# CCA on eclairs objects
+.fastcca <- function(ecl.x, ecl.y, k = NULL, svd.method = c("svd", "irlba", "pcaone") ){
+
+  # checks
+  stopifnot("k must be positive" = k > 0)
+  svd.method <- match.arg(svd.method)
 
   # shrink singular values
   gamma_x <- with(ecl.x, sqrt(dSq/((1-lambda)*dSq + lambda*nu)))
@@ -130,41 +236,24 @@ fastcca <- function(X, Y, k = NULL, k.x=min(dim(X)), k.y=min(dim(Y)), lambda.x=N
   # but latent variables are???
   ax <- with(ecl.x, 1/sqrt(dSq * (1 - lambda) + lambda * nu))
   x.coefs <- dmult(ecl.x$U, ax, "right") %*% dcmp$u
-  rownames(x.coefs) <- colnames(X)
   colnames(x.coefs) <- paste0("comp_", seq(k))
 
   ay <- with(ecl.y, 1/sqrt(dSq * (1 - lambda) + lambda * nu))
   y.coefs <- dmult(ecl.y$U, ay, "right") %*% dcmp$v
-  rownames(y.coefs) <- colnames(Y)
   colnames(y.coefs) <- paste0("comp_", seq(k))
-
-  # Compute latent variables
-  x.vars <- X_scaled %*% x.coefs
-  y.vars <- Y_scaled %*% y.coefs
-
-  # rho <- diag(cor(x.vars, y.vars))[seq(k)]
-  # Faster way to eval diag of correlation
-  rho <- colsums(.standardise(x.vars) * .standardise(y.vars)) / (nrow(X)-1)
-  names(rho) <- paste("comp", seq(k), sep = "")
 
   rho.mod <- dcmp$d
 
   res <- list( 
-        dims = c(n=nrow(X), p1 = ncol(X), p2 = ncol(Y)),
+        dims = c(n=ecl.x$n, p1 = ecl.x$p, p2 = ecl.y$p),
         n.comp = k,
         rho.mod = rho.mod,
-        cor = rho, 
+        # cor = rho, 
         # Cramer's V-statistic for CCA
         cramer.V = sqrt(mean(rho.mod^2)),
         lambdas = c(x = ecl.x$lambda, y = ecl.y$lambda),
         x.coefs = x.coefs, 
-        y.coefs = y.coefs, 
-        x.vars = x.vars, 
-        y.vars = y.vars,
-        x.mu = attr(X_scaled, "mu"),
-        y.mu = attr(Y_scaled, "mu"),
-        x.sd = attr(X_scaled, "sd"),
-        y.sd = attr(Y_scaled, "sd"))
+        y.coefs = y.coefs)
 
   new("fastcca", res)
 }
@@ -222,10 +311,12 @@ predict_from_X <- function( fit, X ){
   # check that column names of X are the same as in 
   # the training data
   if( ! identical(rownames(fit$x.coefs), colnames(X)) ){
-    stop("column names of X doesn't the training data ")
+    stop("column names of X doesn't match the training data ")
   }
 
-  X <- as.matrix(X)
+  if( ! is.matrix(X) ){
+    X <- as.matrix(X)
+  }
 
   # scale input data
   X_scaled <- eachrow(X, fit$x.mu, "-")
@@ -253,10 +344,12 @@ predict_from_Y <- function( fit, Y ){
   # check that column names of Y are the same as in 
   # the training data
   if( ! identical(rownames(fit$y.coefs), colnames(Y)) ){
-    stop("column names of Y doesn't the training data ")
-  }
+    stop("column names of Y doesn't match the training data ")
+  }  
 
-  Y <- as.matrix(Y)
+  if( ! is.matrix(Y) ){
+    Y <- as.matrix(Y)
+  }
 
   # scale input data
   Y_scaled <- eachrow(Y, fit$y.mu, "-")